Reject code files in additionalFiles
additionalFiles previously bypassed every analysis stage — dead-code
elimination, the dependency scanner, and the side-effect classifier all
treated each entry as opaque required content. For non-code assets
(LICENSE, README, JSON data, templates) that's correct, but for a
shipped .js/.ts/.d.ts it meant silent gaps that the readme had to
list as a known limitation.

`combineAllBundleFiles` now refuses any additionalFiles entry whose
targetFilePath matches a code extension, with an error pointing at the
specific path and suggesting the .txt-extension workaround for the
genuine code-as-data cases (templates, code-generator fixtures). The
isCodeFile predicate moves to source/common so both the resource
resolver and the dead-code eliminator can share it.

eliminator.ts was over the 10-import dependency limit after picking up
isCodeFile, so loadBundle and its helpers extract into a focused
load-bundle.ts. The split also clarifies the eliminator's responsibility:
load-bundle.ts owns project/source-file/binding loading and local
reachability; eliminator.ts owns the seeded reachability and transform
orchestration.

The "Known limitations" section is dropped from the readme; the
additionalFiles entry there documents the rejection.
