+++ source/common/code-files.ts
@@ -1,0 +1,5 @@
+const codeFilePattern = /(?:\.d\.ts|\.[cm]?[jt]sx?)$/;
+
+export function isCodeFile(targetFilePath: string): boolean {
+    return codeFilePattern.test(targetFilePath);
+}
