+++ source/dead-code-eliminator/eliminator.ts
@@ -1,94 +1,18 @@
-import { Project, type SourceFile } from 'ts-morph';
-import type { LinkedBundle, LinkedBundleResource } from '../linker/linked-bundle.ts';
+import type { SourceFile } from 'ts-morph';
 import {
     createEmptyFileAnalysis,
     type AnalyzedBundle,
     type AnalyzedBundleResource,
     type DeadCodeEliminator,
-    type EliminationInput,
     type FileAnalysis
 } from './analyzed-bundle.ts';
 import { buildCrossBundleSeeds, type CrossBundleInput } from './cross-bundle/cross-bundle-seeds.ts';
-import { extractTopLevelBindings, type BindingDescriptor } from './reachability/binding-extractor.ts';
-import { bindingId, computeReachability, type FileBindings } from './reachability/reachability.ts';
+import { loadBundle, type LoadedBundle, type LoadedCodeResource, type LoadedResource } from './load-bundle.ts';
+import { bindingId, computeReachability } from './reachability/reachability.ts';
 import { classifySideEffects } from './side-effect-classifier.ts';
-import { computeSideEffectsField, isCodeFile } from './side-effects-field.ts';
+import { computeSideEffectsField } from './side-effects-field.ts';
 import { applyRemovalPlan, type PositionAtom } from './transform/declaration-remover.ts';
 import { recomposeSourceMap } from './transform/source-map-composer.ts';
-
-function createIsolatedProject(): Project {
-    return new Project({});
-}
-
-type LoadedCodeResource = {
-    readonly resource: LinkedBundleResource;
-    readonly sourceFile: SourceFile;
-    readonly bindings: readonly BindingDescriptor[];
-};
-
-type LoadedNonCodeResource = {
-    readonly resource: LinkedBundleResource;
-    readonly sourceFile?: undefined;
-};
-
-type LoadedResource = LoadedCodeResource | LoadedNonCodeResource;
-
-type LoadedBundle = {
-    readonly input: EliminationInput;
-    readonly project: Project;
-    readonly loaded: readonly LoadedResource[];
-    readonly fileBindings: readonly FileBindings[];
-    readonly localReachable: ReadonlySet<string>;
-};
-
-function loadResource(project: Project, resource: LinkedBundleResource): LoadedResource {
-    if (!isCodeFile(resource.fileDescription.targetFilePath)) {
-        return { resource };
-    }
-    const sourceFile = project.createSourceFile(
-        resource.fileDescription.sourceFilePath,
-        resource.fileDescription.content
-    );
-    return { resource, sourceFile, bindings: extractTopLevelBindings(sourceFile) };
-}
-
-function entryPointFilePaths(bundle: LinkedBundle): ReadonlySet<string> {
-    const paths = new Set<string>();
-    for (const entryPoint of bundle.entryPoints) {
-        paths.add(entryPoint.js.sourceFilePath);
-        if (entryPoint.declarationFile !== undefined) {
-            paths.add(entryPoint.declarationFile.sourceFilePath);
-        }
-    }
-    return paths;
-}
-
-function buildFileBindings(loaded: readonly LoadedResource[]): readonly FileBindings[] {
-    const result: FileBindings[] = [];
-    for (const entry of loaded) {
-        if (entry.sourceFile !== undefined) {
-            result.push({
-                sourceFilePath: entry.resource.fileDescription.sourceFilePath,
-                sourceFile: entry.sourceFile,
-                bindings: entry.bindings
-            });
-        }
-    }
-    return result;
-}
-
-function loadBundle(input: EliminationInput): LoadedBundle {
-    const project = createIsolatedProject();
-    const loaded = input.bundle.contents.map((resource) => {
-        return loadResource(project, resource);
-    });
-    const fileBindings = buildFileBindings(loaded);
-    const { reachable: localReachable } = computeReachability({
-        files: fileBindings,
-        entryPointFilePaths: entryPointFilePaths(input.bundle)
-    });
-    return { input, project, loaded, fileBindings, localReachable };
-}
 
 function allBindingNamesFor(loaded: LoadedCodeResource): ReadonlySet<string> {
     const names = new Set<string>();
@@ -228,7 +152,7 @@ function buildMapPathTransformIndex(outputs: readonly BuildOutput[]): ReadonlyMa
 function analyzeBundleWithSeeds(loaded: LoadedBundle, externalSeeds: ReadonlySet<string> | undefined): AnalyzedBundle {
     const { reachable } = computeReachability({
         files: loaded.fileBindings,
-        entryPointFilePaths: entryPointFilePaths(loaded.input.bundle),
+        entryPointFilePaths: loaded.entryPointFilePaths,
         externalSeeds
     });
     const context: AnalysisContext = {
